refactor: extract `eachFilename`
